requirements.txt
matplotlib==3.10.3
nltk==3.9.1
numpy==2.2.5
pandas==2.2.3
python-dotenv==1.1.0
Requests==2.32.3
scikit_learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
streamlit==1.45.0